requirements.txt
phidata  
streamlit
yfinance
numpy
requests
plotly
groq
duckduckgo-search
python-dotenv
googlesearch-python
pycountry
packaging
openai